+++ client/src/assets/sass/_form.scss
@@ -1,24 +1,24 @@
 @import './colors.scss';
 .switch-control {
-    label{
+    label {
         cursor: pointer;
     }
 
-    div.row{
+    div.row {
         display: inline-block;
         margin-right: 25px;
         // &:last-child{
         //     margin-right: 0;
         // }
 
-        input[type="checkbox"]{
+        input[type='checkbox'] {
             display: none;
         }
 
-        input[type="checkbox"]:checked + .switch {
+        input[type='checkbox']:checked + .switch {
             background: $green-color;
 
-            &:before{
+            &:before {
                 -webkit-transform: translateX(30px);
                 -ms-transform: translateX(30px);
                 transform: translateX(30px);
@@ -34,50 +34,48 @@
             background: #000;
             border-radius: 15px;
             margin-right: 13px;
-            transition: all .4s ease-in-out;
+            transition: all 0.4s ease-in-out;
             vertical-align: middle;
 
-            &:before{
+            &:before {
                 content: '';
                 position: absolute;
-                box-shadow: 0 2px 4px 0 rgba(0,0,0,0.5);
-                left:0;
+                box-shadow: 0 2px 4px 0 rgba(0, 0, 0, 0.5);
+                left: 0;
                 top: 0;
                 width: 20px;
                 height: 20px;
                 border-radius: 100%;
                 background: #fff;
-                transition: all .4s ease-in-out;
+                transition: all 0.4s ease-in-out;
             }
         }
     }
 }
 
-
-.tag-select{
+.tag-select {
     display: inline-block;
 
-    .tag-options__value-container{
+    .tag-options__value-container {
         padding: 0 8px;
-
     }
 
     .tag-options__control {
         min-height: 33px;
         font-size: 16px;
         color: #fff;
-        background-color: #2C3034 !important;
+        background-color: #2c3034 !important;
         border-radius: 3px;
         border-color: rgba(151, 151, 151, 0.7);
-        box-shadow: 0 2px 4px -1px rgba(0,0,0,0.2), 0 4px 5px 0 rgba(0,0,0,0.14), 0 1px 10px 0 rgba(0,0,0,0.12);
+        box-shadow: 0 2px 4px -1px rgba(0, 0, 0, 0.2), 0 4px 5px 0 rgba(0, 0, 0, 0.14), 0 1px 10px 0 rgba(0, 0, 0, 0.12);
         min-width: 185px;
         cursor: pointer;
-        &:hover{
+        &:hover {
             border-color: rgba(151, 151, 151, 0.7);
         }
     }
 
-    .tag-options__placeholder{
+    .tag-options__placeholder {
         color: #fff !important;
         font-size: 13px;
     }
@@ -86,25 +84,25 @@
         display: none;
     }
 
-    .tag-options__control--is-focused{
+    .tag-options__control--is-focused {
         outline: none;
         border-color: rgba(151, 151, 151, 0.7);
-        box-shadow: none
+        box-shadow: none;
     }
 
-    .tag-options__single-value{
+    .tag-options__single-value {
         color: #fff;
     }
 
-    .tag-options__menu{
-        background-color: #26282D;
+    .tag-options__menu {
+        background-color: #26282d;
         padding: 0 5px 0 5px;
         border: 1px solid rgba(151, 151, 151, 0.7);
         border-radius: 0;
         min-width: 185px;
     }
 
-    .tag-options__option--is-selected{
+    .tag-options__option--is-selected {
         color: #fff;
         background: transparent;
     }
@@ -112,13 +110,14 @@
     .tag-options__option--is-focused {
         color: #fff;
         &.tag-options__option--is-selected {
-            color: #B5B5B5;
+            color: #b5b5b5;
         }
     }
 
     .tag-options__option {
         padding: 0 0 3px 10px;
-        &, &:hover {
+        &,
+        &:hover {
             background: transparent;
         }
     }
@@ -128,11 +127,11 @@
         color: #fff;
     }
 
-    .tag-options__indicator{
-        padding: 4px
+    .tag-options__indicator {
+        padding: 4px;
     }
 
-    svg{
+    svg {
         color: #fff !important;
     }
 
@@ -140,33 +139,34 @@
         top: 23px;
     }
 
-    .select-row{
+    .select-row {
         font-size: 13px;
         color: #fff;
     }
 
-    .tag-options__multi-value{
+    .tag-options__multi-value {
         background-color: transparent;
         color: #fff;
         position: relative;
         padding-left: 13px;
-        background: #23242B;
+        background: #23242b;
         border-radius: 12px;
         padding-right: 3px;
 
-        .tag-options__multi-value__remove{
+        .tag-options__multi-value__remove {
             position: absolute;
             left: 0;
             top: 4px;
-            &:hover{
+            &:hover {
                 background: transparent;
             }
         }
     }
 }
 
-i{
-    &.indicator, &.select {
+i {
+    &.indicator,
+    &.select {
         border-radius: 100%;
         display: inline-block;
     }
@@ -178,35 +178,35 @@ i{
         vertical-align: middle;
     }
 
-    &.select{
+    &.select {
         width: 10px;
         height: 10px;
         margin-right: 4px;
     }
 
-    &.investigate{
-        background: #FFCD00;
+    &.investigate {
+        background: #ffcd00;
     }
 
-    &.not_investigate{
-        background: #7CA8FF;
+    &.not_investigate {
+        background: #7ca8ff;
     }
 
     &.postpone {
-        background: #A042FF
+        background: #a042ff;
     }
-    &.problem{
-        background: #F64242;
+    &.problem {
+        background: #f64242;
     }
-    &.seen{
-        background: #F5FF00
+    &.seen {
+        background: #f5ff00;
     }
 
     &.normal {
-        background: #45F642;
+        background: #45f642;
     }
 
-    &.untagged{
-        background: #C7C7C7;
+    &.untagged {
+        background: #c7c7c7;
     }
 }
