+++ client/src/components/Header/Header.tsx
@@ -5,9 +5,11 @@ import { Link, useLocation } from 'react-router-dom';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faChevronRight, faChevronLeft, faCaretDown } from '@fortawesome/free-solid-svg-icons';
 import { getCurrentExperimentDetails } from 'src/model/selectors/experiment';
+import { filterEventsByTagAction } from 'src/model/actions/datarun';
 import { getSelectedExperiment } from '../../model/selectors/projects';
 import { onUserLogoutAction } from '../../model/actions/users';
 import { RootState } from '../../model/types';
+import Dropdown from '../Common/Dropdown';
 
 type StateProps = ReturnType<typeof mapState>;
 type DispatchProps = ReturnType<typeof mapDispatch>;
@@ -15,7 +17,7 @@ type Props = StateProps & DispatchProps;
 
 export const Header: React.FC<Props> = (props) => {
   const isSwitchVisible = props.selectedExperimentID ? 'active' : '';
-  const { experimentDetails } = props;
+  const { experimentDetails, filterByTags } = props;
   let location = useLocation();
 
   const currentView = location.pathname;
@@ -28,6 +30,13 @@ export const Header: React.FC<Props> = (props) => {
 
   const [isInfoOpen, toggleInfo] = useState(false);
   const activeClass = isInfoOpen ? 'active' : '';
+
+  const dropDownProps = {
+    isMulti: true,
+    closeMenuOnSelect: false,
+    placeholder: 'Filter',
+    onChange: filterByTags,
+  };
 
   window.addEventListener('click', (evt: Event) => {
     if (currentView === '/') {
@@ -48,43 +57,51 @@ export const Header: React.FC<Props> = (props) => {
           <FontAwesomeIcon icon={currentView === '/' ? faChevronRight : faChevronLeft} />
         </Link>
         {currentView !== '/' && (
-          <div className="exp-info" onClick={() => toggleInfo(!isInfoOpen)}>
-            <ul>
-              <li className={activeClass}>
-                Details
-                <span>
-                  <FontAwesomeIcon icon={faCaretDown} />
-                </span>
-                <ul>
-                  <li>
-                    <span>Pipeline:</span>
-                    <span>{experimentDetails.pipeline}</span>
-                  </li>
-                  <li>
-                    <span>Dataset:</span>
-                    <span>{experimentDetails.dataset}</span>
-                  </li>
-                  <li>
-                    <span>By:</span>
-                    <span>{experimentDetails.created_by || 'Unknown'}</span>
-                  </li>
-                  <li>
-                    <span>Project:</span>
-                    <span>{experimentDetails.project}</span>
-                  </li>
-                </ul>
-              </li>
-            </ul>
+          <div className="header-left-wrapper">
+            <div className="exp-info" onClick={() => toggleInfo(!isInfoOpen)}>
+              <ul>
+                <li className={activeClass}>
+                  Details
+                  <span>
+                    <FontAwesomeIcon icon={faCaretDown} />
+                  </span>
+                  <ul>
+                    <li>
+                      <span>Pipeline:</span>
+                      <span>{experimentDetails.pipeline}</span>
+                    </li>
+                    <li>
+                      <span>Dataset:</span>
+                      <span>{experimentDetails.dataset}</span>
+                    </li>
+                    <li>
+                      <span>By:</span>
+                      <span>{experimentDetails.created_by || 'Unknown'}</span>
+                    </li>
+                    <li>
+                      <span>Project:</span>
+                      <span>{experimentDetails.project}</span>
+                    </li>
+                  </ul>
+                </li>
+              </ul>
+            </div>
+            <div className="tag-wrapper">
+              <Dropdown {...dropDownProps} />
+            </div>
           </div>
         )}
 
-        <ul className="user-opts">
-          <li>
-            <button type="button" onClick={logoutUser} className="logout-button">
-              Logout
-            </button>
-          </li>
-        </ul>
+        <div className="header-right-wrapper">
+          <ul className="user-opts">
+            <li>
+              <button type="button" onClick={logoutUser} className="logout-button">
+                Logout
+              </button>
+            </li>
+          </ul>
+          <div className="clear" />
+        </div>
       </div>
     </header>
   );
@@ -97,6 +114,7 @@ const mapState = (state: RootState) => ({
 
 const mapDispatch = (dispatch: Function) => ({
   userLogout: () => dispatch(onUserLogoutAction()),
+  filterByTags: (tags) => dispatch(filterEventsByTagAction(tags)),
 });
 
 export default connect<StateProps, DispatchProps, {}, RootState>(mapState, mapDispatch)(Header);
